# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Innovation Circle API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendees:
    post:
      operationId: createAttendee
      tags: [attendees]
      summary: Create an attendee (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendee"
        "400":
          description: Invalid attendee or Hub
        "409":
          description: Email already exists
    get:
      operationId: listAttendees
      tags: [attendees]
      summary: List attendees (admin sees all, attendee sees their circle)
      parameters:
        - name: circleId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of attendees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendeeWithActivity"
  /attendees/import:
    post:
      operationId: importAttendees
      tags: [attendees]
      summary: Import attendees from CSV rows (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendeeImportInput"
      responses:
        "200":
          description: Import completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendeeImportResult"
        "400":
          description: Invalid import payload or Hub
        "403":
          description: Admin access required
  /attendees/{id}:
    get:
      operationId: getAttendee
      tags: [attendees]
      summary: Get attendee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attendee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendeeWithActivity"
        "404":
          description: Not found
    patch:
      operationId: updateAttendee
      tags: [attendees]
      summary: Update attendee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendee"

  # Meetings
...
  /admin/send-reminder:
    post:
      operationId: sendPreMeetingReminder
      tags: [admin]
      summary: Email pre-meeting reminder with open goals to all attendees
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminEmailTrigger"
      responses:
        "200":
          description: Reminder sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/send-survey:
    post:
      operationId: sendPostMeetingSurvey
      tags: [admin]
      summary: Email post-meeting CSAT survey to all attendees
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminEmailTrigger"
      responses:
        "200":
          description: Survey emails sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Get admin dashboard overview
      parameters:
        - name: circleId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboard"

  # Storage
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Auth
...
    Attendee:
      type: object
      required: [id, name, email, company, role, circleId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        company:
          type: string
        role:
          type: string
        circleId:
          type: integer
        createdAt:
          type: string
    AttendeeInput:
      type: object
      required: [name, email, circleId]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        company:
          type: string
        circleId:
          type: integer
    AttendeeImportRow:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        company:
          type: string
    AttendeeImportInput:
      type: object
      required: [circleId, attendees]
      properties:
        circleId:
          type: integer
        attendees:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: "#/components/schemas/AttendeeImportRow"
    AttendeeImportSkippedRow:
      type: object
      required: [row, email, reason]
      properties:
        row:
          type: integer
        email:
          type: string
        reason:
          type: string
          enum: [duplicate_file, duplicate_existing]
    AttendeeImportResult:
      type: object
      required: [createdCount, skippedCount, created, skipped]
      properties:
        createdCount:
          type: integer
        skippedCount:
          type: integer
        created:
          type: array
          items:
            $ref: "#/components/schemas/Attendee"
        skipped:
          type: array
          items:
            $ref: "#/components/schemas/AttendeeImportSkippedRow"
    AttendeeWithActivity:
      type: object
      required: [id, name, email, company, role, circleId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        company:
          type: string
        role:
          type: string
        circleId:
          type: integer
        createdAt:
          type: string
        lastActivityAt:
          type: ["string", "null"]
        goalCount:
          type: integer
        surveyResponseCount:
          type: integer
    AttendeeUpdate:
      type: object
      properties:
        name:
          type: string
        company:
          type: string

    # Meeting
...
    AdminEmailTrigger:
      type: object
      required: [circleId]
      properties:
        circleId:
          type: integer
        meetingId:
          type: integer
    AdminDashboard:
      type: object
      required: [totalAttendees, totalMeetings, totalGoals, goalsByStatus, recentActivity]
      properties:
        totalAttendees:
          type: integer
        totalMeetings:
          type: integer
        totalGoals:
          type: integer
        goalsByStatus:
          type: object
          properties:
            new:
              type: integer
            inProgress:
              type: integer
            completed:
              type: integer
            notStarted:
              type: integer
        pendingInvites:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
    ActivityItem:
      type: object
      required: [type, description, createdAt]
      properties:
        type:
          type: string
        description:
          type: string
        createdAt:
          type: string
        attendeeName:
          type: ["string", "null"]

    # Storage
...